Update build_config.h with more recent Chromium version
The goal is to make library compilable on all platforms supported by
Debian. From this point of view is more reliable to rely on a project
which is already building all required platforms.

The initial issue was non-detectable endianess for little endian PPC
system. Possibly due to typo. Using more battle-tested code seems to
be better way to go.

This change technically breaks compatibility with PPC64 family symbols
(the define changed its name). This define is not used by the library
itself and the build_config.h is not intended to be used by users of
this library.